削除機能の条件を追加

+++ app/controllers/items_controller.rb
@@ -2,7 +2,6 @@ class ItemsController < ApplicationController
   before_action :authenticate_user!, only: [:new, :edit, :destroy]
   before_action :set_item, only: [:show, :edit, :update]
   before_action :move_to_index, only: :edit
-
   def index
     @items = Item.includes(:user).order('created_at DESC')
   end
@@ -36,8 +35,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item.user == current_user
+      item.destroy
+      redirect_to root_path
+    else
+      render :index
+    end
   end
 
   private
